Refactor code_folder to src across files
