+++ docs/Code.rst
@@ -5,81 +5,81 @@ Code documentation
 Initialization
 --------------
 
-.. automodule:: code_folder.A_initialization
+.. automodule:: src.A_initialization
    :members:
    :undoc-members:
 
 Data input
 ----------
 
-.. automodule:: code_folder.B0_data_input_json
+.. automodule:: src.B0_data_input_json
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.B2_read_csv
+.. automodule:: src.B2_read_csv
    :members:
    :undoc-members:
 
 Data processing
 ---------------
 
-.. automodule:: code_folder.C0_data_processing
+.. automodule:: src.C0_data_processing
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.C1_verification
+.. automodule:: src.C1_verification
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.C2_economic_functions
+.. automodule:: src.C2_economic_functions
    :members:
    :undoc-members:
 
 Modelling
 ---------
 
-.. automodule:: code_folder.D0_modelling_and_optimization
+.. automodule:: src.D0_modelling_and_optimization
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.D1_model_components
+.. automodule:: src.D1_model_components
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.D2_model_constraints
+.. automodule:: src.D2_model_constraints
    :members:
    :undoc-members:
 
 Evaluation
 ----------
 
-.. automodule:: code_folder.E0_evaluation
+.. automodule:: src.E0_evaluation
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.E1_process_results
+.. automodule:: src.E1_process_results
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.E2_economics
+.. automodule:: src.E2_economics
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.E3_indicator_calculation
+.. automodule:: src.E3_indicator_calculation
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.E4_verification_of_constraints
+.. automodule:: src.E4_verification_of_constraints
    :members:
    :undoc-members:
 
 Output
 ------
 
-.. automodule:: code_folder.F0_output
+.. automodule:: src.F0_output
    :members:
    :undoc-members:
 
-.. automodule:: code_folder.F1_plotting
+.. automodule:: src.F1_plotting
    :members:
    :undoc-members:
