+++ src/D0_modelling_and_optimization.py
@@ -7,7 +7,7 @@ import pprint as pp
 try:
     from .D1_model_components import define_oemof_component, call_component, helpers
 except ImportError:
-    from code_folder.D1_model_components import (
+    from src.D1_model_components import (
         define_oemof_component,
         call_component,
         helpers,
